requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
redis==5.0.1
eventlet==0.35.1
gunicorn==21.2.0
PyJWT==2.8.0
Werkzeug==3.0.1